+++ src/index.js
@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom';
 import React from 'react';
-import App from './App'
+import App from './components/App'
 
  
  ReactDOM.render(<App />, document.getElementById('app'));
